add README

+++ src/com/speical/first/Pro222.java
@@ -2,18 +2,20 @@ package com.speical.first;
 
 import java.util.Scanner;
 /**
- * N阶台阶问题
+ * N阶台阶问题（典型动态规划）
  * 
  * kinds[i] = kinds[i - 1] + kinds[i - 2]
  * 意思就是走到i台阶步数等于走到i-1的步数加上走到i-2的步数
  * 因为走到i台阶只能由这两种台阶情况构成
+ * 
+ * 题目给的数小,如果再大点，此题就上升为动态规划+大数相加
  * @author Special
  * @time 2018/02/11 20:06:54
  */
 public class Pro222 {
 	
 	static int SIZE = 90 + 5;
-	static int[] kinds = new int[SIZE];
+	static long[] kinds = new long[SIZE];
 	
 	public static void init() {
 		kinds[0] = kinds[1] = 1;
